+++ tests/testthat/helper-rlang.R
@@ -9,3 +9,16 @@ zap_srcref_attributes <- function(x) {
   attr(x, "wholeSrcref") <- NULL
   x
 }
+
+run_script <- function(file, envvars = chr()) {
+  skip_on_os("windows")
+
+  # Suppress non-zero exit warnings
+  suppressWarnings(system2(
+    file.path(R.home("bin"), "Rscript"),
+    c("--vanilla", file),
+    stdout = TRUE,
+    stderr = TRUE,
+    env = envvars
+  ))
+}